moved router from index to app lol
